#优化代码

+++ app/src/main/java/com/cxb/myfamilytree/widget/familytree/FamilyDBHelper.java
@@ -81,7 +81,7 @@ public class FamilyDBHelper {
                 .where(sql, parentId, parentId, ignoreChildId, birthday));
     }
 
-    public long save(List<FamilyBean> families) {
+    public int save(List<FamilyBean> families) {
         return liteOrm.save(families);
     }
